# Add the performance test program