This program measures the runtime for the compositing functions. We can observe the runtime change with the compositing size.

For example,
```c
do_all_tests("Pixmap", 7, 18, 2, 1000000);
```
the result is as follows:
```bash
[ Pixmap: 7x7x1000000 ]
argb32_source_argb32                     0.045891 sec
argb32_over_argb32                       0.075607 sec
[ Pixmap: 9x9x1000000 ]
argb32_source_argb32                     0.068147 sec
argb32_over_argb32                       0.111900 sec
[ Pixmap: 11x11x1000000 ]
argb32_source_argb32                     0.086605 sec
argb32_over_argb32                       0.151332 sec
[ Pixmap: 13x13x1000000 ]
argb32_source_argb32                     0.110934 sec
argb32_over_argb32                       0.203429 sec
[ Pixmap: 15x15x1000000 ]
argb32_source_argb32                     0.133251 sec
argb32_over_argb32                       0.255378 sec
[ Pixmap: 17x17x1000000 ]
argb32_source_argb32                     0.157086 sec
argb32_over_argb32                       0.314955 sec
```