@jserv 2024-10-25
Consider the following changes for simplicity:
```diff
--- a/Makefile
+++ b/Makefile
@@ -112,15 +112,13 @@ endif
 
 # Performance tester
 
-ifeq ($(CONFIG_PERFORMANCE_TEST), y)
-target-$(CONFIG_PERFORMANCE_TEST) += apps/mado-perf
-apps/mado-perf_depends-y += $(target.a-y)
-apps/mado-perf_files-y += apps/perf.c
-apps/mado-perf_includes-y := include
-apps/mado-perf_ldflags-y := \
+target-$(CONFIG_PERFORMANCE_TEST) += mado-perf
+mado-perf_depends-y += $(target.a-y)
+mado-perf_files-y += apps/perf.c
+mado-perf_includes-y := include
+mado-perf_ldflags-y := \
        $(target.a-y) \
        $(TARGET_LIBS)
-endif
 
 # Standalone application
 
--- a/configs/Kconfig
+++ b/configs/Kconfig
@@ -58,16 +58,12 @@ config LOADER_GIF
 
 endmenu
 
-menu "Performance Testers"
+menu "Applications"
 
 config PERFORMANCE_TEST
-    bool "Build composite tester"
+    bool "Build performance tester"
     default y
 
-endmenu
-
-menu "Demo Applications"
-
 config DEMO_APPLICATIONS
     bool "Build demo applications"
     default y
@@ -106,4 +102,4 @@ config DEMO_ANIMATION
     bool "Build animation demo"
     default y
     depends on DEMO_APPLICATIONS 
```