@jserv 2024-11-03
Take the result of [x11perf](https://www.x.org/archive/X11R7.5/doc/man/man1/x11perf.1.html) for reference:
```
The X.Org Foundation server version 12302006 on :0
Sun Nov  3 17:41:37 2024

Sync time adjustment is 4.2587 msecs.

           40000000 reps @   0.0000 msec (177000000.0/sec): 1x1 rectangle
   40000000 reps @   0.0000 msec (178000000.0/sec): 1x1 rectangle
   40000000 reps @   0.0000 msec (177000000.0/sec): 1x1 rectangle
   40000000 reps @   0.0000 msec (178000000.0/sec): 1x1 rectangle
   40000000 reps @   0.0000 msec (176000000.0/sec): 1x1 rectangle
  200000000 trep @   0.0000 msec (177000000.0/sec): 1x1 rectangle

  200000000 reps @   0.0000 msec (31300000.0/sec): 10x10 rectangle
  200000000 reps @   0.0000 msec (34700000.0/sec): 10x10 rectangle
  200000000 reps @   0.0000 msec (36200000.0/sec): 10x10 rectangle
  200000000 reps @   0.0000 msec (37300000.0/sec): 10x10 rectangle
  200000000 reps @   0.0000 msec (30100000.0/sec): 10x10 rectangle
 1000000000 trep @   0.0000 msec (33700000.0/sec): 10x10 rectangle
...
```